inserimento metodo:inserisci prodotto

+++ ApplicationCore/src/main/java/com/howtodoinjava/app/applicationcore/repository/ClienteRepository.java
@@ -1,8 +1,9 @@
 package com.howtodoinjava.app.applicationcore.repository;
+
 import com.howtodoinjava.app.applicationcore.entity.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+public interface ClienteRepository extends JpaRepository<Cliente, Long>{
 }
